requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.10
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.4.4
click==8.1.6
coloredlogs==15.0.1
confection==0.1.1
ctransformers==0.2.18
cymem==2.0.7
cytoolz==0.12.2
dataclasses-json==0.5.14
decorator==5.1.1
diskcache==5.6.1

fastapi==0.99.1
filelock==3.12.2
Flask==2.3.2
flatbuffers==23.5.26
floret==0.10.4
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
hnswlib==0.7.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
itsdangerous==2.1.2
jellyfish==1.0.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
langchain==0.0.250
langcodes==3.3.0
langsmith==0.0.16

markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.2
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pathy==0.10.2
Pillow==9.5.0
pip==23.2.1
posthog==3.0.1
preshed==3.0.8
protobuf==4.23.4
pulsar-client==3.2.0
py-cpuinfo==9.0.0

pydantic==1.10.12
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
pyphen==0.14.0
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==68.0.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.19
srsly==2.4.7
starlette==0.27.0
streamlit==1.25.0
streamlit-chat==0.1.1
streamlit-elements==0.1.0
streamlit-on-Hover-tabs==1.0.1
streamlit-option-menu==0.3.6
sympy==1.12
tenacity==8.2.2
textacy==0.13.0
thinc==8.1.11
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.31.0
typer==0.9.0
typing_extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
validators==0.20.0
wasabi==1.1.2
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.7
wheel==0.41.0
yarl==1.9.2
zipp==3.16.2
